move references out to shared DLL's for services and models
