+++ agileways.usermgt.admin.client/Server/Controllers/DirectoryObjects/AssignmentsController.cs
@@ -2,7 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using agileways.usermgt.shared.Services;
+using agileways.usermgt.shared.server.Services;
 using agileways.usermgt.shared.Models.DirectoryObjects;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
