+++ agileways.usermgt.admin.client/Server/Controllers/DirectoryObjects/ServicePrincipalController.cs
@@ -1,7 +1,7 @@
 
 
 using System.Threading.Tasks;
-using agileways.usermgt.shared.Services;
+using agileways.usermgt.shared.server.Services;
 using agileways.usermgt.shared.Models.DirectoryObjects;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
