--- agileways.usermgt.admin.client/Shared/DirectoryObjects/AppRoleAssignedTo.cs
@@ -1,17 +1,0 @@
-
-
-namespace agileways.usermgt.admin.client.Shared.DirectoryObjects
-{
-
-    public class AppRoleAssignedTo
-    {
-        public string Id { get; set; }
-        public string PrincipalDisplayName { get; set; }
-        public string ResourceDisplayName { get; set; }
-        public string AppRoleId { get; set; }
-        public string AppRoleName { get; set; }
-        public string ResourceId { get; set; }
-        public string PrincipalId { get; set; }
-        public string PrincipalType { get; set; }
-    }
-}
