--- agileways.usermgt.admin.client/Shared/DirectoryObjects/AppRoleAssignment.cs
@@ -1,12 +1,0 @@
-
-
-namespace agileways.usermgt.admin.client.Shared.DirectoryObjects
-{
-
-    public class AppRoleAssignments
-    {
-        public string PrincipalDisplayName { get; set; }
-        public string ResourceDisplayName { get; set; }
-
-    }
-}
