+++ agileways.usermgt.api/Controllers/UserController.cs
@@ -1,8 +1,8 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using agileways.usermgt.api.Services;
-using agileways.usermgt.api.Shared.DirectoryObjects;
+using agileways.usermgt.shared.server.Services;
+using agileways.usermgt.shared.Models.DirectoryObjects;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
