--- agileways.usermgt.api/Shared/DirectoryObjects/AppRoleAssignedTo.cs
@@ -1,16 +1,0 @@
-
-
-namespace agileways.usermgt.api.Shared.DirectoryObjects
-{
-
-    public class AppRoleAssignedTo
-    {
-        public string Id { get; set; }
-        public string PrincipalDisplayName { get; set; }
-        public string ResourceDisplayName { get; set; }
-        public string AppRoleId { get; set; }
-        public string ResourceId { get; set; }
-        public string PrincipalId { get; set; }
-        public string PrincipalType { get; set; }
-    }
-}
