--- agileways.usermgt.api/Shared/DirectoryObjects/AppRoleAssignment.cs
@@ -1,12 +1,0 @@
-
-
-namespace agileways.usermgt.api.Shared.DirectoryObjects
-{
-
-    public class AppRoleAssignments
-    {
-        public string PrincipalDisplayName { get; set; }
-        public string ResourceDisplayName { get; set; }
-
-    }
-}
