--- agileways.usermgt.api/Shared/DirectoryObjects/Role.cs
@@ -1,11 +1,0 @@
-
-namespace agileways.usermgt.api.Shared.DirectoryObjects
-{
-    public class Role
-    {
-        public string Id { get; set; }
-        public string Name { get; set; }
-        public string Value { get; set; }
-        public string Description { get; set; }
-    }
-}
