--- agileways.usermgt.api/Shared/DirectoryObjects/User.cs
@@ -1,16 +1,0 @@
-
-using System.Collections.Generic;
-
-namespace agileways.usermgt.api.Shared.DirectoryObjects
-{
-    public class User
-    {
-        public string DisplayName { get; set; }
-        public string GivenName { get; set; }
-        public string Surname { get; set; }
-        public string Id { get; set; }
-        public string UserPrincipalName { get; set; }
-        public IEnumerable<UserIdentity> Identities { get; set; }
-        public string CompanyId { get; set; }
-    }
-}
