--- agileways.usermgt.api/Shared/DirectoryObjects/UserIdentity.cs
@@ -1,12 +1,0 @@
-
-
-namespace agileways.usermgt.api.Shared.DirectoryObjects
-{
-
-    public class UserIdentity
-    {
-        public string SignInType { get; set; }
-        public string Issuer { get; set; }
-        public string IssuerAssignedId { get; set; }
-    }
-}
