+++ agileways.usermgt.api/agileways.usermgt.api.csproj
@@ -12,4 +12,9 @@
     <PackageReference Include="Microsoft.Identity.Web.UI" Version="1.1.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\agileways.usermgt.shared\agileways.usermgt.shared.csproj" />
+    <ProjectReference Include="..\agileways.usermgt.shared.server\agileways.usermgt.shared.server.csproj" />
+  </ItemGroup>
+
 </Project>
