+++ agileways.usermgt.shared/Models/DirectoryObjects/User.cs
@@ -11,5 +11,6 @@ namespace agileways.usermgt.shared.Models.DirectoryObjects
         public string Id { get; set; }
         public string UserPrincipalName { get; set; }
         public IEnumerable<UserIdentity> Identities { get; set; }
+        public string CompanyId { get; set; }
     }
 }
