+++ agileways.usermgt.shared/agileways.usermgt.shared.csproj
@@ -4,12 +4,4 @@
     <TargetFramework>net5.0</TargetFramework>
   </PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="5.0.6" />
-    <PackageReference Include="Microsoft.Authentication.WebAssembly.Msal" Version="5.0.6" />
-    <PackageReference Include="Microsoft.Graph" Version="3.33.0" />
-    <PackageReference Include="Microsoft.Graph.Auth" Version="1.0.0-preview.7" />
-    <PackageReference Include="Microsoft.Identity.Web" Version="1.1.0" />
-  </ItemGroup>
-
 </Project>
